refactor(codex): split goal API docs and client
