+++ ap-web/src/components/codex/index.ts
@@ -1,4 +1,4 @@
 export { CodexGoalControl, CodexGoalStatusPill } from "./CodexGoalControl";
 export { CodexGoalDialog } from "./CodexGoalDialog";
 export { useCodexGoalState } from "./useCodexGoalState";
-export type { CodexGoal } from "@/lib/sessionsApi";
+export type { CodexGoal } from "@/lib/codexGoalApi";
